+++ src/System.Text.Json/tests/System.Text.Json.Tests.csproj
@@ -45,6 +45,8 @@
     <Compile Include="Serialization\TestClasses.SimpleTestClass.cs" />
     <Compile Include="Serialization\TestClasses.SimpleTestClassWithNullables.cs" />
     <Compile Include="Serialization\TestClasses.SimpleTestClassWithObject.cs" />
+    <Compile Include="Serialization\TestClasses.SimpleTestClassWithObjectArrays.cs" />
+    <Compile Include="Serialization\TestClasses.SimpleTestClassWithSimpleObject.cs" />
     <Compile Include="Serialization\TestData.cs" />
     <Compile Include="Serialization\Value.ReadTests.cs" />
     <Compile Include="Serialization\Value.WriteTests.cs" />
